+++ src/main/java/net/darmo_creations/mccode/interpreter/builtin_functions/CeilFunction.java
@@ -22,6 +22,6 @@ public class CeilFunction extends BuiltinFunction {
 
   @Override
   public Object apply(final Scope scope) {
-    return (int) Math.ceil(this.getParameterValue(scope, 0));
+    return (long) Math.ceil(this.getParameterValue(scope, 0));
   }
 }
