+++ src/main/java/net/darmo_creations/mccode/interpreter/builtin_functions/MinFunction.java
@@ -39,7 +39,7 @@ public class MinFunction extends BuiltinFunction {
     } else if (p instanceof Range) {
       Range r = (Range) p;
       return StreamSupport.stream(Spliterators.spliteratorUnknownSize(r.iterator(), 0), false)
-          .min(Integer::compareTo).orElseThrow(() -> new EmptyCollectionException(scope));
+          .min(Long::compareTo).orElseThrow(() -> new EmptyCollectionException(scope));
     }
     throw new CastException(scope, ProgramManager.getTypeInstance(ListType.class), ProgramManager.getTypeForValue(p));
   }
